requirements.txt
apify-client==1.6.1
asyncpraw==7.8.0
bittensor==8.5.0
jupyter==1.0.0
numpy==2.0.1
pydantic==2.10.1
python-dotenv==1.0.0
pytz==2023.3.post1
rich==13.7.0
torch==2.5.1
wandb==0.18.7
pandas==2.2.2
cryptography==43.0.3
requests==2.32.3
huggingface-hub==0.23.3
datasets==2.20.0
pyarrow==17.0.0
psutil==5.9.8
loguru==0.7.3